Add loading cache infrastructure
This is loosely based on Guava's CacheBuilder, although much less
versatile and tunable.

Items are reference-counted and evicted after a fixed amount of time
after they've been created, although the policy to evict after the last
time might be another good candidate.

Since this is a loading cache, it knows how to create and destroy the
entries; this greatly simplifies the interface (one just asks the cache
for an item referring to a key, and if it's not there, it's automatically
created).

Read-write locks from pthread are used in this implementation, but this
might move to something that scales better, such as RCU or similar
technique.

This isn't used right now, but the idea is to have this be transparent
for all the handlers.  Also, this haven't been extensively tested.  We
might hit a few bugs whenever lwan switches to this infrastructure, but
you're not relying on lwan, are you?