Plug memory leak after ae1ee8bf80
If a connection is aborted because of some reason, a request's string
buffer won't be freed. Using deferred statements to free the string
buffer won't work on success paths as it will be out of scope at that
time.

This effectively locks the coro_defer() behavior of stacking
statements; a comment was added to make sure future refactor keep this
in mind.