Add preliminary Lua module
This has a lot of room for improvement, both in performance and feature
set.  The idea is to actually make it work first, then slowly improve
it.  This can maybe close #13.

For instance:

- Although LuaJIT is used if found, the script is reparsed every time,
giving no chance for any kind of optimization to actually happen.

- A script file can only contain a single endpoint (unless glue code in
Lua is written to deal with dispatch, etc).  Ideally functions could be
defined so that a single script file could handle multiple endpoints;
for instance, to handle "GET /lua/hello" to a module that answers to
prefix "/lua", one would define the following function:

	function handle_get_hello(request, response)
		name = request.query_string["name"]
		if name then
			response.say("Hello, " .. name .. "!")
		else
			response.say("Hello, World!")
		end
	end

- Only "lwan.say()" is implemented. This maps to "strbuf_set()"
followed by "lwan_response_send_chunk()".