Add request pipelining support.
The way this works is simple: instead of discarding a buffer
potentially full of requests after sending a response to one, lookahead
a few bytes after the last request terminator and see if it matches a
known HTTP verb.  If it does, it is another request; move everything
from that terminator to the beginning of the buffer and continue
processing as if nothing happened (eventually reading more from the
request socket if needed).

Some numbers, using a pipelined script with wrk (testing on localhost
here, which is far from ideal, but the only possibility I have right
now):

	Running 10s test @ http://127.0.0.1:8080/plaintext
	  2 threads and 100 connections
	  Thread Stats   Avg      Stdev     Max   +/- Stdev
	    Latency    10.52ms    7.03ms  33.87ms   57.95%
	    Req/Sec   195.33k    18.31k  288.05k    71.70%
	  3834734 requests in 10.02s, 3.51GB read
	  Non-2xx or 3xx responses: 2586005
	Requests/sec: 382644.71
	Transfer/sec:    359.15MB

And without the pipelined script:

	Running 10s test @ http://127.0.0.1:8080/plaintext
	  2 threads and 100 connections
	  Thread Stats   Avg      Stdev     Max   +/- Stdev
	    Latency   257.90us  487.09us  23.74ms   99.29%
	    Req/Sec   122.80k     9.94k  142.22k    88.99%
	  2308700 requests in 10.00s, 521.81MB read
	Requests/sec: 230892.17
	Transfer/sec:     52.19MB

Latency goes higher when attending to pipelined requests, but
throughput is significantly higher.