Ensure coroutines resumes on socket read readiness on buffer exhaustion
This patch essentially fixes the pipelining support added with commit
7f191c9.  It introduces the connection flag CONN_MUST_READ, which
forces the I/O loop to switch the epoll watch to a "can read" status,
regardless of what kind of events a particular connection is interested
in at the moment.  This way, if the request buffer is empty, a
coroutine will only be resumed if its associated connection is ready to
be read from.

The code in `resume_coro_if_needed()` is still pretty confusing,
specially regarding the meaning of the `write_events` variable and
`CONN_WRITE_EVENTS` flag.  I've been postponing the rewrite of this
part for a long time, and will eventually do this when I finally
implement timers.

Also, I'm not happy with the memmove() call in `read_request()`. The
request buffer should probably be something else, such as a ring
buffer, and if a request is set to be pipelined, epoll should watch for
input events as well, and fill this buffer (i.e. the scheduler will
probably have to be rewritten, but for a good reason) as soon as data
is available from the socket.