Fix build where pthread_barrier_t is unavailable
This should be fixed differently. I'll possibly have to resort to GCC's
`#include_next` (which is also supported by Clang) and create a missing
directory, with header names the same as the header names in Linux, and
pass that directory first.

Also, structs that are in the missing.h can't be used in public headers,
since missing.h won't be installed.  This can be sort of worked around
by creating a `lwan-build-options.h` with all the `-DHAS_*` flags passed
to GCC, and installing that.

Another attempt to help with #160.