Use smaller chunks when sending data with sendfile()
The rationale is that, if the server is very busy, one fast connection
shouldn't dominate the whole worker thread.

Instead of calling sendfile() with the file size and try to send it in
one go, send the first 128KiB first (or less, if the file is smaller
than that), then send in chunks of 512KiB (or less).  Since each chunk
yields the coroutine, this should give readahead() some time to actually
finish populating the page cache.

This is far from the ideal solution, of course. Ideally, Linux
sendfile() should have a flag similar to the FreeBSD version,
controlling if it should block or not waiting for the disk.  But since
that's not there and it's unlikely to ever be there, make do.