Store module information in ELF/Mach-O sections
Instead of relying on dlopen() to work on executable files, store
module information in a special ELF/Mach-O section ("lwan_module" and
"__DATA,lwan_module" respectively) instead.

This simplifies how modules are looked up, and reduces the possibility
of a misconfiguration calling a function that it shouldn't call.

This also removes the module registry hash table, since looking for
modules is a pretty simple linear search on all structs defined in the
special section.  This is most likely quicker than the hash table
lookup, considering that there are very few modules in Lwan; in any
case, since module lookup is performed only once during program boot,
this should have never been a concern.

Confirmed working on FreeBSD, macOS, and Linux.