Simplify declaration of HTTP status tables
Use a single array for both the short status (e.g. "404 Not Found") and
the descriptive error message (e.g.  "The requested resource could not
be found by this server.").

Apart from adding a new enumeration value, this is now the only other
place to modify when adding a new response code.