Do not link with the math library unless necessary
Some dependencies (such as Lua) might pull in -lm, but if Lwan is built
without it, this should reduce its size a tiny wee little bit.  This
uses a builtin GCC function (__builtin_fpclassify), which is used by
the GNU C library to implement fpclassify(), that is the only use of
the math library in Lwan.  This builtin is also implemented by Clang.