Fix potential race condition when scheduling new connection
Commit 24cb2952 introduced a race condition that could manifest when an
I/O thread would be unblocked in the call to epoll_wait() before
handling the eventfd nudge, effectively trying to resume a coroutine
that doesn't exist.

Move the epoll_ctl(EPOLL_CTL_ADD, conn_fd) call to the accept_nudge()
function, and call it only as it consumes file descriptors from the
queue.