Allow requests to sleep
Use a timer wheel to keep track of timers, and use the epoll timeout as
a way to turn the wheel.  Epoll will wait for at most 1000ms each time,
and the timeout value will be controlled by the wheel.  The death queue
is still controlling the connection lifetime, so a coroutine can't wait
for more than the keep_alive_timeout configuration parameter.