Remove `ended` member from struct coro
This was introduced to avoid a memcpy() in the last context switch of a
coroutine.  However, most coroutines that matter (i.e.  coroutines
handling requests) will be freed before ended would be true, as that
function never returns.

The only case where that memcpy() would be avoided would be in
templates with sequences; however, that's not a hot path, so this
optimization has no point.

Saves 4 bytes per coroutine (due to alignment), and a (hopefully
well-predicted) branch every context switch, as the memcpy() to the
switcher is always necessary.