Not being able to raise number of file descriptors isn't an error
This is required to use pledge() on OpenBSD: without relaxing the
pledge too much by enabling the "proc" promise, which would also enable
fork(2) -- which is definitely not wanted -- make it a warning instead.

So, instead of being a critical error, just use system defaults from
limits.h.  Also, if for some reason the system does not provide
OPEN_MAX, use the oldschool NOFILE from <sys/param.h>.  (I have not
tested on all OSes supported by Lwan if this header is available
everywhere.  If it fails, the build system will have to figure it out.)