Move calls to readahead() to a low priority thread
This thread has SCHED_IDLE priority, and, if on Linux, will also have
IDLE I/O class.  The idea is that readahead should never have any kind
of impact on any of the worker threads (which can't block, as they
perform cooperative context switches).

The file serving module, that is now making use of this, will call this
function every time it adds a file to its cache.  Items in cache are
alive for a few seconds only, so this will play well in keeping the
operating system cache fresh.

Threads can request a file descriptor to be readahead by calling
lwan_readahead_queue(fd, size); this writes the request to a pipe,
which is consumed by the readahead thread.  Only the read side of that
pipe is blocking; if the writer side can't write the readahead command,
it's not the end of the world: this is just an optimization.