Error events shouldn't be ignored while coalescing kevents
When watching events that aren't EPOLLIN or EPOLLOUT (e.g. only to keep
the file descriptor in the epoll set to be notified of errors when a
coroutine is suspended by a timer), udata is set to 1.  When coalescing
the events, udata should only be checked if the filter is EVFILT_WRITE;
ignoring anything that happened if udata is 1, regardless of the filter
in the kevent, will actually ignore EV_ERROR and EV_EOF events.