Port jwz's Dali Clock to Lwan
This is a quick and dirty port from the Pebble port of Dali Clock,
which was based off of Jamie Zawinski's famous original, that has
been porteed to inumerous platforms[1].

I think this is the first time this clock has been ported to the
Web without JavaScript, though.

Have fun!

[1] https://www.jwz.org/xdaliclock/