Automatically resize hash table buckets array
Use a simple (and maybe not optimal) heuristic to either duplicate,
or reduce by half, the buckets array, when items are respectively
added or removed from the hash table.  This heuristic can be tuned
later.

While new buckets for entries will be found during the rehashing
process, there's no need to calculate the hash again during
resizing, since the hash value is kept in the hash_entry struct.