Store mask, rather than number of buckets
When the number of buckets was constant, "hash & (constant - 1)" would
be compiled down to two operations, with "- 1" being calculated by the
compiler.  When this was moved to a member in a struct, the "- 1" was
being calculated during runtime.

Store only the mask (size_as_power_of_two - 1), as masking the least
significant bits of a hash is a common operation.

Also, put some asserts in place to ensure that the bucket sizes will
remain a power of two.