Put NUL terminator in the correct place in coro_strndup()
The NUL terminator was being written always at the last byte in a chunk
of memory of max_len bytes.  This might not correspond to the size of
the string, if the string to be copied is smaller than that buffer.

Use strnlen() to calculate the minimum between the max desired length
and the actual size for the string being copied.

To avoid strlen() being called in addition to strnlen() when
coro_strdup() (which is implemented on top of coro_strndup()) is
called, always pass `SIZE_MAX - 1` as the maximum length instead.  This
is as safe as calling strlen() anyway.