Refactor header parsing code
This cleans up a lot of the code for parsing headers, making it more
idiomatic and reducing amount of repeated code due to extensive use of
macros.

By splitting the loop in two, this not only opens the opportunity to
use SIMD to find the start and end pointers for every header, this also
makes it possible to reduce calls to memchr() and the likes, making the
parser also faster in the end.  This also removes calls to unbounded
string search functions, and gets rid of some undefined behavior.

Also, by using compound statements, the implicit assignment of value
and length is removed, making the code a lot easier to read and
maintain.

All tests are passing, even those for pipelined requests.

The only drawback is that the current version limits each request to
32 headers.