Simplify parsing of Connection header
Only one character is saved for this header: no need to allocate a struct
lwan_value temporarily to obtain the first character from the header just
to throw it away immediately after.  Provide a HEADER_RAW macro that does
most of the processing of HEADER, but returns the pointer to the header
value rather than a struct lwan_value.  Then define HEADER in terms of
HEADER_RAW.