Make it possible to watch file descriptors
The idea here is to re-use the thread queue to signal when a coroutine
should be resumed: writing a negative file descriptor will signal that,
if the coroutine responsible for the absolute value is suspended, it
should be resumed instead.  This will allow requests to sleep until
there's activity on a particular file descriptor.

Since the thread queue is SPSC, only the main thread is capable of
watching these file descriptors.  As a result, it now contains its own
epoll.  To keep things idiomatic, instead of passing callbacks for each
event type, coroutines are used instead: the result of coro_yield() are
the events reported back by epoll_wait().

This interface is still considered private, hence only declared in
lwan-private.h.