Go back to using level-triggered epoll events
This dramatically reduces the amount of system calls necessary to handle
requests.  When running testrunner, making 1M requests with 1k clients,
this table shows the number of epoll_ctl() calls:

                             /hello             /chunked
   With this patch:            1005                 2005
        Previously:        20000005              2000132

The same approach was problematic with pipelined requests; however,
this version fixes that by flipping the flags (from EPOLLIN to
EPOLLOUT) when there's a potential request in the pipeline, forcing the
connection coroutine to be resumed.

The hello world benchmark is ~15% faster.