Headers were being ignored after first request in the same connection
The parse_headers() refactor introduced a bug, where the headers for a
request would only be parsed if it were the first request made in a
connection in the cases where multiple packets had to be sent to
compose a request (say, if manually making a request through nc): a
request would be fulfilled right after receiving the verb, path, and
HTTP version, rather than wait for either an empty line, or the
headers.

Fix this by checking if there's anything in the pipeline other than the
terminating NUL byte, that's always written by the
read_from_request_socket() function.

Also, change the signature from `char *process_request(...)` to
`bool process_request(...)`, and return `true` if all headers could
be parsed, and `false` iff the `headers_start` array overflows.