Consider integer overflow when building string in rewrite module
Although user input is passed to append_str(), it's never going to be
larger than the whole HTTP request, and that must fit within the stack
of a coroutine, so the possibility of overflowing a 64-bit size_t are
practically impossible.

While not technically necessary in this case, it's cheap to check for
overflow, so do it even if for clarity alone.