Add libFuzzer driver for the request parsing code
To enable, build with -DBUILD_FUZZER=ON and CC environment variable set
to clang.  No samples will be built, liblwan will be instrumented, and
the driver will be available as src/bin/fuzz/fuzz.

Right now this fuzzes only with address sanitizer, but there are plans
to build a fuzzing binary for each sanitizer that makes sense.

This fuzzes only the request parsing code.  There are plans to add
fuzzing drivers for other parts of the code, including the rewriting
engine, configuration file reader, template parser, and URL routing.

A good dictionary and corpus to use are from the H2O web server,
available in its repository, `fuzz' directory[1].  I'm planning to add
a corpus specific to Lwan in the future, possibly getting from the
test suite, as it contains tests to exercise corner cases that have
been fixed in the past.

Bugs found with this fuzzer have been already fixed.  They were using
an older version of the request parser fuzzing function that didn't
consider that the request should be fully complete to be parsed.  This
has been since fixed in this version.

[1] https://github.com/h2o/h2o