Retry/Complete sendfile when incomplete.
Set CONN_FLIP_FLAGS in lwan_sendfile if sendfile has written less bytes than
requested and sendfile needs to be called again for the remaining bytes. This
ensures that epoll_ctl will be called to set EPOLLOUT and we can return to the
coroutine and write the remaining bytes after epoll_pwait returns the next
time.

Fixes commit 0318fc5.