Reduce amount of memory necessary to build response headers
At least in mod-serve-files.  The header buffers were being allocated
with DEFAULT_BUFFER_SIZE, which by default is 4096 bytes, but the size
passed to prepare_headers() was always DEFAULT_HEADERS_SIZE, which is
512 by default.

This reduces the possibility of running out of coroutine stack space.

Clean up how the header buffer is passed to prepare_headers(), assuming
that a buffer of size `static DEFAULT_HEADERS_SIZE` will be passed and
not requiring a size_t to be passed to that function.  Since this is
an static function and buffers are always passed with a fixed size,
this is safe.