Fix read overflow while parsing request headers
This happens when the first four bytes of a request header is matched
near the end of a request, but the ": " substring would be beyond the
buffer.

Fix this by always checking if there's enough space for the full header
name from the current position.  As a side effect, check for the
end-of-request marker just once while looking for header matches,
without needing to store one empty pair of start/end-of-header
pointers.

Credit to OSS-Fuzz: https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=14154