Reduce indirection when looking up extension in mime_type table
Instead of processing data so that there's an array of pointers
pointing to the decompressed buffer, store it differently.  Previously,
data was stored as a string of extension-NUL-mime-type-NUL; after
decompression, two arrays were populated, containing pointers to the
data.  It now contains all extensions, padded with NUL bytes (from 0 to
7 NULs), then followed by all the MIME types, separated by NUL; after
decompression, Lwan spends time populating only an array of pointers
to MIME types.

With this change, the cache lines that would be used for the extension
pointers are now used for the data itself.  This is 120 fewer cache
lines ((sizeof(char*) * 954) / 64) that need to be, in the worst case,
populated in order to find an extension.

As a side-effect, all 9 extensions from mime.types that are over 8
characters long (which aren't common), are not included in the MIME
type table any longer.  Not a bit deal.  As such, the binary size
decreased a few hundred bytes due to having fewer elements, and
the compressor liking all the NUL padding.

When compressing with Brotli, compressed table now sits at 6480 bytes,
from 6739 before this change.  Uncompressed data is larger due to
padding, but is offset from not needing the pointer array, so it's
a win-win scenario.

This could probably be improved further by using something like gperf,
though, but, considering that this information is cached by the file
serving module, it's not really a hot path.