Temporary cache entry keys need to be freed when unreffing
The only exception is when they ended up being in the hash table but
trying to acquire the queue lock failed for some reason.  Otherwise,
since entries are never added to the hash table, hash_del() will never
call free() on the key, so do it in cache_entry_unref().