Remove one malloc roundtrip per connection
This makes it possible to define array types that have a "inline first"
allocation strategy, where data is initially allocated inline with the
array type, and, if needed, move on to heap allocation.  The size of
this inline storage is 16 elements, the same number of elements that
triggers a reallocation when using heap-allocated storage.

This is now used for the deferred callbacks arrays.  Most handlers
defers at most around 5 callbacks, so there's plenty of room in the
happy path.  Since all connections will have at least one deferred
callback (to free the strbuf), this removes a malloc roundtrip for all
connections.