Fix coroutine yield_value signaling
When moving the coro_defer array to inlinefirst, the size of that structure
changed, and with it the offset of the yield_value member was pushed
forward.  Since, on x86_64, the trampoline function is written in assembly,
this is was corrupting memory.

(A static_assert() is now in place to ensure that this doesn't happen again.)