Provide a bump-pointer allocator for coro_malloc()
This will create "bump pointer arenas" that are at most 1024 bytes
long.  (This value has been chosen arbitrarily and might change in the
future, but this should be transparent to coro_malloc() users.)

If there's enough space in the arena after aligning the size to a
sizeof(void*) boundary, allocation is super quick: just a pointer
increment.  If the requested size is greater than 1024, malloc() will
be used instead, as usual; if there's not enough space in the current
arena (and it fits in 1024 bytes), a new one will be created.  Arenas
are tracked with coro_defer, which is already efficient as they're
stored in a inlinefirst array.

Most uses of coro_malloc() are through coro_strdup(), which are to copy
small strings.  This is important in benchmarks such as "Fortunes" from
TechEmpower Web Framework benchmarks, which Lwan is participating
again.

A few other places in Lwan will use coro_strdup(), including Lua
support and the rewrite module.  coro_malloc() is used by a few things,
including the HTTP authorization module.

Unfortunately it's not possible to use this for coro_malloc_full(),
which assumes that the user-provided callback will be responsible for
freeing the memory.

BPA is disabled when building Lwan for fuzzing, and if Valgrind support
is enabled, the new regions are notified through client requests (much
like coroutine stacks are).  ASAN support will be implemented later
(it's simple to do, just have to do it).