Restore "capacity" field in strbuf
Without this field, when the request coroutine resets the response
string buffer, this will trigger a reallocation next time the string
buffer is used, even though the allocation is (probably) big enough to
contain the new data that will be stored in that strbuf.

This is in preparation for the TechEmpower Web Frameworks benchmark,
more specifically for the JSON and plaintext benchmarks which send the
same pre-prepared response over and over again.