Remove some branches when encoding JSON objects and array
In both cases, the branches were there to avoid generating a trailing
`,' after the last item.

For objects, take advantage of the fact that the element order doesn't
matter, so loop through the 2nd to the last key/value pair first,
always appending a comma to the output, then append the first key/value
pair.

For arrays, loop through n_elements-1, always adding the comma, and
then write the next element.