Reduce calls to append_bytes() when encoding JSON objects
By giving a hint to the JSON encoder that the keys do not need to be
encoded (e.g.  because we know beforehand that they don't require any
special treatment), we can reduce calls to append_bytes() in hot paths.

The TWFB benchmark implements append_bytes() by append a character at a
time to a strbuf, which is inneficient: it's writing two bytes at a
time (the desired char + terminating NUL), and updating the internal
bookkeeping data every time for every character.  If there's no need to
encode the key, just a single memory copy+bookkeeping information is
needed for each key.

In the "queries" benchmark, if requesting for 500 queries, this reduces
the number of calls to append_bytes() from 13001 to 7001.  This is
almost 12KiB of data per request that does not need to be written
byte-per-byte.