Allow a connection coroutine to wait on other file descriptors
This sets up epoll to wait on an arbitrary file descriptor using
EPOLLONESHOT; useful, for instance, if implementing asynchronous
database access using the MariaDB client libraries in non-blocking
mode.

A sample program that proxies a TCP connection (and sends back the
bytes received as chunks) to 127.0.0.1:6969 is provided to
demonstrate that this works.