Remove fd_watcher
With the async await stuff, this is not needed anymore.  This
simplifies the main accept loop: instead of blocking on epoll_wait(),
it now blocks on accept4() -- which becomes non-blocking while a herd
is passing, to maintain the behavior of nudging the worker threads only
when the herd is gone.  Without a coroutine to handle the incoming
connections, the main loop should be a tiny wee little bit more
efficient now.