Remove most branches while serializing JSON in the TWFB benchmark
The likelyhood of append_bytes() failing is very minimal, and the
benchmark doesn't care about which particular error caused the
serialization to fail; it only cares if it failed.

So, instead of branching for each and every call to see if it failed,
perform a bitwise OR with the return value of every call.  This way, if
all calls succeed (i.e. return value is 0), the overall serialization
value is also success (0).  Otherwise, it's going to be a bitwise OR of
all negative errno values returned by the user-provided append_bytes()
function, which may or may not be useful (if they return always the
same value, or if they return something different than a negative errno
-- e.g. one bit set per kind of error).

Since this is mostly for benchmarking purposes, this optimization is
perfectly fine.