Add preliminary SignalR Chat demo
Inspired by demo by David Fowler and Damian Edwards in this
presentation: https://www.youtube.com/watch?v=iL9nLAjCPtM

This is not yet hooked up to the build system, and I don't even know if
this builds.  There are lots of issues with the WebSockets API in Lwan
that I need to fix before this can become a nice demo.