lwan_sendfile() fallback: resolve FIXME
As noted in the FIXME,

    /* FIXME: is this correct?  fd being read here is a file, not
     * a socket, so a WANT_READ may actually lead to a coro never
     * being woken up */

This problem is indeed triggered when running the test_range_no_to()
test on OpenBSD (where lwan uses this fallback implementation); lwan
reads until end-of-file, then times out waiting for the file to become
ready to read.

Resolve this by trying to read the file again, synchronously.

The test_range_no_to() testcase is one of the few testcases where the
data to be sent is not a multiple of the buffer size (i.e. of 512).