Separate coroutine and its stack in debug builds and OpenBSD
Due to an OpenBSD requirement, b698d334 started allocating a coroutine
and its stack using mmap() directly.  This patch changes this, so that
the stack isn't allocated inline with the coroutine struct anymore, but
separately; this adds a little bit of robustness as it's harder for
stack overflows to corrupt a coroutine struct (and possibly reuse the
indirect jump in the context switch as an ROP gadget).

Also use this behavior by default in the debug builds, to catch issues
that the stack_poison member was supposed to be used for, but without
relying on fragile userland instrumentation.