Avoid re-arming async/await watchers every time I/O is requested
By avoiding the use of EPOLLONESHOT, we can avoid calling epoll_ctl()
to re-arm the watcher every time a read or a write to a file descriptor
is requested.

lwan_request_async_write() and lwan_request_async_read() will use,
respectively, send() and recv(), passing the MSG_DONTWAIT flag, so that
if they return -EWOULDBLOCK, then a request to epoll to wake the coroutine
is put in place.

This saves not only a roundtrip to the kernel every time one wants to
read/write a async fd, but also one coroutine context swap.