Avoid using coro_defer() in the fast path for the cached_world benchmark
Instead of using cache_coro_get_and_ref_entry(), define a local
my_cache_coro_get_and_ref_entry() function that does not schedule a
unref when the coroutine is done.

This avoids demoting the coro_defer array to the heap and, for large
number of queries, avoids a large number of calls to realloc().

This change is fine because between reading the cached entry and
copying it to the queries_json array the coroutine can't be destroyed
from the outside.r