Allow peeking at a WebSocket connection to write an event pump
If a handler now tries to read from a WebSocket connection, it'll
return values different than true/false:

  - ENOTCONN: read() called before connection was upgraded
  - ECONNRESET: Client has closed the connection
  - EAGAIN: Nothing is available, try again
  - 0: There's something in the request buffer, go read it!

This way, it's possible to write a message pump by looking at the
return value of this function.  At the time, there's no possibilty
of performing a read with a given timeout, so if a handler sleeps
for a while, that sleep won't be canceled.  This will eventually
be fixed, but it should now be possible to write event pumps with
this and write chat pub/sub applications.

The example has been modified as well to showcase the possibilities,
by implementing an echo server that counts the time since the last
message that was received.

It's still all barebones, and I'm not happy with the API still, but
it's coming along.

This change also fixes a lot of stuff when reading WebSocket frames
and is more standards compliant.