Fix race condition in pubsub while publishing
Incrementing the refcount of a shared message has to be done in a
atomic way; taking sub->lock isn't protecting this shared resource,
and another subscriber for this message might be consuming it as
that message is being published to other subscribers.

By initializaing the refcount to 1 when the message is allocated,
atomically increasing the number, and dropping the message when
we atomically decrease the reference  count after the message has
been published guarantees that the refcount won't be corrupted.

Also, this fixes a memory leak when dropping the published message
when there were no subscribers: the message itself was not being
freed.