Reduce memory allocation churn in pubsub
Use a doubly-linked list of ring buffers to store pointers to
struct lwan_pubsub_msg, instead of allocating a linked list node
per message in the queue.

All subscription queues begin with a single node in the linked list.
If that overflows, a new one is created (and added to the list tail)
and the message is put there.

Messages are consumed from each linked list node in succession; if
one of these nodes contain an empty ring buffer, the entire node
is removed from the linked list, and messages are then consumed from
the ring buffer in the next node.