Implement private function to extend a strbuf
This is inherently unsafe because it changes the "used" field in the
strbuf as well; it's supposed to be used when the string buffer is
being used to do I/O and the allocated buffer is going to be read()
into.

It's not being exposed in lwan-strbuf.h, only in lwan-private.h. It
also has a "_unsafe" prefix in the function name, to be explicit
about how dangerous this function can be.