Fix memory leaks when requesting query_param, cookies, or post_data
This was tricky to find.  Since the storage for the helper struct lived in
the frame for lwan_process_coro(), once it returned, all the coro_defer()
calls were trying to free up the arrays whose storage was in the dead call
frame.  I have no clue how this all worked until now!

Moved the helper struct to the coroutine, where the lifetime of that struct
is guaranteed to exist while the coroutine is cleaning up its things.

This should also fix a performance regression found with the TWFB preparations
for the round 20, where Lwan didn't fare well in the cached queries benchmark,
where it's one of the places it should really shine.

Fixes #283.