Cleanup read_from_request_socket() and finalizer function prototype
This renames read_from_request_socket() to client_read(), and gets rid
of the "total_read" parameter (which can be inferred from buffer).

It also has some other minor changes, including always yielding with
CONN_CORO_WANT_READ (for both EINTR and EAGAIN), renaming of
FINALIZER_ERROR_TIMEOUT to FINALIZER_TIMEOUT, and the usage of a string
switch to find the header terminator in the fast path.