Fix reading of POST request bodies when needing second read()
When issuing a second read() call to finish reading the request body, a
buffer is created in the stack and passed to client_read().  The
finalizer function called by client_read() would assume that the
helper->buffer struct would be the same as the one created in the frame
that called client_read().  Since the number of bytes read would never
coincide with the amount of bytes we wanted to read, the whole
operation would time out.

Re-introduce the buffer parameter to the finalizer function, so that we
can be sure that the buffer it operates on is the one we're interested
in.

This is fine as this a internal-only API and changing won't break
anybody.  (Hopefully.)

These tests were not passing locally but were failing in the CI system.