Add application/wasm Web Assembly mime type
This adds the application/wasm mime type, which is used to represent
compiled Web Assembly files, and registers it with the associated file
extension, .wasm.

While this isn't a registered standard, it's a widely recognized
mimetype, and is the one officially set out by the WASM committee and
current JavaScript standards.

It's concretely useful to use this mimetype, as it enables use of the
"WebAssembly.instantiateStreaming" browser API.

References:

https://webassembly.github.io/spec/core/binary/conventions.html
 - the webassembly document recommending application/wasm

https://github.com/WebAssembly/spec/issues/573
 - progress on getting this registered as an official mimetype

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/WebAssembly/instantiateStreaming#Instantiating_streaming
 - JS API which requires the application/wasm mimetype to function