Log requests in lwan-request.c, not lwan-response.c
This is the more logical place to log them: when processing a request,
and not a response.  The only problem here is, however, that until
url_map->handler() returns, we don't know if it succeeded or not if
it's something like an infinite loop, say, in a websockets server
handler.