Provide better support for libucontext
For some reason, the Makefile in libucontext is broken -- building
it with meson fixes the generated libraries, which do not export
the symbols and as a result, no symbol interposition ends up taking
place.

To force those symbols to be used, use the `libucontext_` prefixed
versions; at least we'll get a build error in Lwan if libucontext
wasn't built correctly rather than silently building and then
using the libc ucontext functions and definitions.

I have confirmed that Lwan doesn't make calls to block/unblock the
signal mask on AARCH64, which happens when using the libc ucontext
functions.  This should improve performance on non-x86 systems
where libucontext is supported.