Trim request string buffer between requests/websocket messages
While this was fine for the usual request side of things, a WebSockets
client could potentially uncontrollably make Lwan allocate memory until
malloc would exhaust because the string buffer never trimmed the
allocated buffer (it only changed the buffer length, never its
capacity/allocated buffer).

Fix this by introducing a new lwan_strbuf_reset_trim() API, which works
pretty much like lwan_strbuf_reset(), but takes a size_t parameter that
forces the old buffer to freed and a new buffer to be allocated with
that size *if* the old capacity exceeds the parameter.

With this, each request will now trim the string buffer to 2KB between
requests in the same connection, and, by default, each WebSockets
message read will force the buffer to be 1KB or smaller before
potentially reallocating; for WebSockets, though, it's possible to tune
the hint parameter if the expected packet should be larger than 1KB.