If read_request() fails, just abort the coroutine after responding
If read_request() returns any error at this point, it's probably better
to just send an error response and abort the coroutine and let the
client handle the error instead: we don't have information to even log
the request because it has not been parsed yet at this stage.

Even if there are other requests waiting in the pipeline, this seems
like the safer thing to do.