backup() can't be called by the parser
This would only work if the lexer's ring buffer is empty; otherwise,
tokens would be produced in the wrong order.  Remove the call and
extract the code that handles a closing bracket to its own function, to
act as the next state from both parse_config() and parse_key_value().