Make all threads accept connections from clients
This moves the responsibility of accepting connections from the main
thread to the various worker threads that are spawned to handle
connections.  A single accepting thread was a bottleneck and would
compete with busy worker threads; now all worker threads share this
responsibility.

Requires, for the moment, SO_REUSEPORT support; support for systems
where this is disabled or not available will come later.