Fix timers after getting rid of eventfd in b8231695
Previous to the change in b8231695, a thread's epoll would wake up
from a nudge, have all coroutines for new connections created, client
file descriptors added to the epoll set, and then it would wait for
epoll_wait to notify of events to handle.

With the change, coroutines would be resumed right after being created,
and would not give an opportunity to turn the timer wheel. So requests
that happened right after the server started would consider the amount
of time that had passed since the thread boot time to the sleep time.

Move to the previous behavior by not resuming the coroutine off the
bat, and give an opportunity for the timer wheel to turn.