Connections may be dropped before coroutine was created for it
A connection might receive a EPOLLHUP|EPOLLRDHUP event before we spawn
a coroutine to handle that client.  Even though the connection has been
reset by the peer, the file descriptor is still alive.  Since
timeout_queue_expire() will only be called whenever we have a file
descriptor already, we can always close the connection, regardless if
there's a coroutine or not.

This has not been tested. I also do not remember why that check was
there in the first place, but this is very old and part of the initial
implementation of the timeout queue.  This might need to be reversed in
the future if we re-discover the reason it was like that in the first
place.