Remove need to set SO_INCOMING_CPU when accepting clients
When I originally wrote the code to use a CBPF program to direct the
incoming sockets to the right CPU, it didn't seem to work as intended,
especially if one used getsockopt(SO_INCOMING_CPU) to inspect the value
and compare it against what thread->cpu gives you.

The issue, that I realized only after reading [1], is that listening
sockets must be created in a particular order if it is to use
SKF_AD_CPU as the value returned by the CBPF program.

I also realized while reading the same blog post, that my CBPF program
was wrong, because it was MOD-ing the CPU ID with the total number of
sockets, which is obviously wrong in retrospect.  I also fixed this
to only extract the SKF_AD_CPU field as I should've since the
beginning.

With this patch, threads are created serially, in the correct order.
Calling getsockopt(SO_INCOMING_CPU) in a pinned worker thread will
return the expected value, that coincides with thread->cpu and what
getcpu() says.  Because of this, setsockopt(SO_INCOMING_CPU) for each
accepted socket isn't called any longer.

[1] https://talawah.io/blog/extreme-http-performance-tuning-one-point-two-million/