If strbuf:set_static() is called, buffer doesn't have fixed size anymore
BUFFER_FIXED should only be set if strbuf has been initialized with a
fixed size; if the buffer is changed (e.g. by calling the set_static()
method), then that flag doesn't make any sense -- the buffer is passed
as a constant, and as such it makes no sense for strbuf to try to
grow that buffer in any operation.