Fix TLS connection handshake issue
This was a race condition that sometimes happened when a thread set the
connection flags to request a TLS setup after adding the file
descriptor to another thread's epoll set.  If the request were accepted
by the other thread's I/O thread, the CONN_NEEDS_TLS_SETUP flag wasn't
yet set, so Lwan never initiated the TLS handshake.

Fix this by setting the flag before adding the file descriptor to the
epoll set.