Reduce number of TCP packets by half for TLS handshakes
Use our own I/O wrapper functions instead of the ones provided by
mbedTLS so that MSG_MORE is used to bundle parts of the handshake that
are sent by the server.

This is still not perfect, as we're don't know when mbedTLS is going to
switch from "only sending stuff" to "let's read some stuff", so we
flush the buffer by setting the corking bit.  We have to do this twice.
(An alternative would be counting the number of packets, but that is a
bit brittle.)