Add preliminary FastCGI support
This is sufficient to use Lwan as a proxy to PHP-FPM.  I have not
tested with other FastCGI implementations, but I suspect it'll work
just fine.

Right now, this is about 70% as fast as nginx in the same machine.
From a cursory glance at strace, it seems that nginx achieves this by
reading as much as possible from the socket instead of reading each
record every time it goes back to the processing loop like Lwan does.
Maybe someday we can optimize this further, but I wanted to get this
working version out first.  In any case, I'm quite happy with this
result, especially since I have given optimization not a lot of
thought yet.

I have not fuzz-tested or done any rigorous code review, so this might
not be that safe.  I'm tired and it's possible something trivial
slipped.  I have to think how I'm going to exercise this code with a
fuzzer so I can hook it up to oss-fuzz or something.

Some parameters are still hardcoded.  I'm not sure if the ones that
aren't will be sufficient for most applications.

POST requests (or any other requests that will require a request body)
won't work, as the STDIN record isn't implemented yet.