Address out-of-bounds read in huffman decoder for HTTP/2
Since this code isn't being used at the moment, this is perfectly fine.
It just means that the fuzzer is doing its job.

This still needs to be optimized later, but let's not worry about that
for now.  Let's just leave the fuzzer do its thing and find all the
corner cases we'll need to think about when it's actually time to use
this.

Credits to oss-fuzz:
    https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=46506