Bubble awaited socket hangup event to request handler
The way async/await works in Lwan is that, when registering the
awaited fd with epoll, the data pointer still points to the
HTTP client connection conn.  This ensures that the same mechanism
that's used for all connections is used to resume them.  This
also has the effect that, if the peer of an awaited fd closes its
connection, we get a hangup event that will close the HTTP
client connection as well.

The behavior as is is wrong, since there are two file descriptors
registered in epoll with the same conn pointer, and we would
call close on both of them, potentially causing a race condition
that could cause an unrelated fd to be closed.

This patch checks if a conn is suspended by looking at two flags:
CONN_SUSPENDED and CONN_HAS_REMOVE_SLEEP_DEFER.  If both are set,
then it means that we're not unsuspending a coro due to async/await,
but rather because it was sleeping with a timer.  By checking if
the former is set but not the latter, we're sure that this is an
awaited fd, so we can skip its expiration and let the request handler
handle the situation.

Request handlers are supposed to defer a callback that closes the
fds it's watching, so this all works out.

Found by running:

    while true; do
         curl http://localhost:8080/php/index.php || break
    done

While running a simple PHP script.  The loop would eventually be
broken. With this fix, this doesn't happen anymore.