Consider prefix when building redir location string
In the file serving module, if one requests, say, "/blog/foo", where
"foo" is a directory, we were redirecting to "blog/foo/" instead, which
would in fact request "/blog/blog/foo/", which isn't what we want.
Consider the prefix (from the configuration file) when building this
string, so that, for instance, if the prefix is "/", we redirect to
"/blog/foo/", which is what we want.